json-xstream bug
git-svn-id: http://geowebcache.org/svn/trunk@327 e7b91dd5-889a-44ae-8e97-0abfc27e49b2

+++ geowebcache/src/main/java/org/geowebcache/seeder/SeedTask.java
@@ -53,8 +53,8 @@ public class SeedTask {
             
             TileLayer layer = RESTDispatcher.getAllLayers().get(req.getName());
             log.info("Begin seeding layer : " + layer.getName());
-            int zoomStart = req.getZoomStart();
-            int zoomStop = req.getZoomStop();
+            int zoomStart = req.getZoomStart().intValue();
+            int zoomStop = req.getZoomStop().intValue();
             MimeType mimeType = null;
             try {
                 mimeType = MimeType.createFromFormat(req.getMimeFormat());
